__app__/requirements.txt
azure-functions
azure-storage-blob
requests
beautifulsoup4
psycopg2
pandas
twilio
tweepy
gspread
oauth2client